Fixed datetime bug and inconsistent whitespace formattting.
Begin time is set and passed in from bash script.  End time (if not passed in) is
defaulted in the python script to the current datetime.  If there is enough of a delay
that the timestamps differ by at least 1 second, it results in an extra (and presumably
unintended) iteration, that tries to retrieve and shred tomorrow's report file.

- Fixed the bug by updating script to compare date values only
- Also converted tabs to spaces since the file was a mix of both, mostly spaces

Outstanding issues (assuming we continue to maintain this script):
- Fix the default 'begintime' to be a numeric timestamp. '-1 days' will cause the script to fail.
- Update ftp logic to check if the file exists.